requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
google-genai==0.3.0
